:bug: async/await works for rendering the main page now
console.log(1.main)
