+++ controllers/imageController.js
@@ -8,27 +8,31 @@ exports.mainPage = async (req, res) => {
 	const dislikes = []
 
 	// find all images to display on start
-	await Image.find({}, (err, images) => {
-		images.forEach((obj) => {
+	await Image.find({}, async (err, images) => {
+		await images.forEach((obj) => {
 			imageArray.push(obj)
 		})
+		console.log('1. main')
 	})
 
 	// find a dislike list to see which image has been rated already
-	await User.find({ user_id: req.session.userId }, (err, user) => {
-		user[0].dislikes.forEach((dislike) => {
+	await User.find({ user_id: req.session.userId }, async (err, user) => {
+		await user[0].dislikes.forEach((dislike) => {
 			dislikes.push(dislike.toString())
 		})
+		console.log('2. main')
 	})
 
-	res.render('main', {
+	// render page
+	await res.render('main', {
 		userName: req.session.userName,
 		images: imageArray,
 		dislikes: dislikes,
 	})
+	console.log('3. main')
 
+	// get a socket connection
 	const io = req.app.get('io')
-
 	io.on('connection', (socket) => {
 		console.log('socket connected!')
 	})
