+++ views/main.ejs
@@ -1,6 +1,5 @@
 <%- include ./part/header.ejs %>
 
-
 	<h1>Welcome <%= userName %></h1>
 
 	<div class="pics">
